+++ graphdot/microkernel/__init__.py
@@ -4,7 +4,7 @@
 Microkernels are positive-semidefinite functions between individual nodes and
 edges of graphs.
 """
-from ._base import MicroKernel, Constant
+from ._base import MicroKernel, Constant, Normalize
 
 from .product import Product
 from .kronecker_delta import KroneckerDelta
@@ -12,13 +12,13 @@ from .kronecker_delta import KroneckerDelta
 from .square_exponential import SquareExponential
 from .rational_quadratic import RationalQuadratic
 
-from .normalize import Normalize
-
 from .composite import Composite
 from .additive import Additive
 from .tensor_product import TensorProduct
 
 from .convolution import Convolution
+
+from .dotproduct import DotProduct
 
 
 __all__ = [
@@ -33,4 +33,5 @@ __all__ = [
     'TensorProduct',
     'Additive',
     'Convolution',
+    'DotProduct'
 ]