+++ graphdot/microkernel/convolution.py
@@ -5,23 +5,26 @@ from graphdot.codegen.cpptool import cpptype
 from graphdot.codegen.template import Template
 from graphdot.util.pretty_tuple import pretty_tuple
 from ._base import MicroKernel
-from .normalize import Normalize
 
 
-def Convolution(kernel: MicroKernel, normalize=True):
-    r"""Creates a convolutional microkernel, which sums up evaluations of a
+def Convolution(kernel: MicroKernel, mean=True):
+    r"""Creates a convolutional microkernel, which averages evaluations of a
     base microkernel between pairs of elements of two variable-length feature
     sequences.
-    :math:`k_{conv}(X, Y) = \sum_{x \in X} \sum_{y \in Y} k_{base}(x, y)`
 
     Parameters
     ----------
     kernel: MicroKernel
         The base kernel can be any elementary or composite microkernels,
         while the attribute to be convolved should be sequences.
-    normalize: bool
-        Whether or not to normalize the convolution result to ensure it is a
-        valid microkernel in range [0, 1].
+    mean: bool
+        If True, return the arithmetic mean of the kernel evaluations, i.e.
+        :math:`k_{conv}(X, Y) =
+        \frac{\sum_{x \in X} \sum_{y \in Y} k_{base}(x, y)}{|X||Y|}`.
+        Otherwise, return the sum of the kernel evaluations, i.e.
+        :math:`k_{conv}(X, Y) = \sum_{x \in X} \sum_{y \in Y} k_{base}(x, y)`.
+        Thus, this serves as a means of normalization beyonds the dot product
+        based one.
     """
 
     @cpptype(kernel=kernel.dtype)
@@ -30,17 +33,24 @@ def Convolution(kernel: MicroKernel, normalize=True):
         def name(self):
             return 'Convolution'
 
-        def __init__(self, kernel):
+        def __init__(self, kernel, mean):
             self.kernel = kernel
+            self.mean = mean
 
         def __call__(self, X, Y, jac=False):
             if jac is True:
                 Fxy, Jxy = list(zip(*[
                     self.kernel(x, y, jac=True) for x in X for y in Y
                 ]))
-                return np.sum(Fxy), np.sum(Jxy, axis=0)
+                if self.mean:
+                    return np.mean(Fxy), np.mean(Jxy, axis=0)
+                else:
+                    return np.sum(Fxy), np.sum(Jxy, axis=0)
             else:
-                return np.sum([self.kernel(x, y) for x in X for y in Y])
+                if self.mean:
+                    return np.mean([self.kernel(x, y) for x in X for y in Y])
+                else:
+                    return np.sum([self.kernel(x, y) for x in X for y in Y])
 
         def __repr__(self):
             return f'{self.name}({repr(self.kernel)})'
@@ -48,17 +58,20 @@ def Convolution(kernel: MicroKernel, normalize=True):
         def gen_expr(self, x, y, theta_scope=''):
             F, J = self.kernel.gen_expr('_1', '_2', theta_scope + 'kernel.')
             f = Template(
-                r'convolution([&](auto _1, auto _2){return ${f};}, ${x}, ${y})'
+                r'''convolution<${mean}>(
+                        [&](auto _1, auto _2){return ${f};}, ${x}, ${y}
+                )'''
             ).render(
-                x=x, y=y, f=F
+                mean='true' if self.mean else 'false', x=x, y=y, f=F
             )
             template = Template(
-                r'''convolution_jacobian(
+                r'''convolution_jacobian<${mean}>(
                         [&](auto _1, auto _2){return ${j};},
                         ${x}, ${y}
                 )'''
             )
-            jacobian = [template.render(x=x, y=y, j=j) for j in J]
+            mean = 'true' if self.mean else 'false'
+            jacobian = [template.render(mean=mean, x=x, y=y, j=j) for j in J]
             return f, jacobian
 
         @property
@@ -76,7 +89,8 @@ def Convolution(kernel: MicroKernel, normalize=True):
         def bounds(self):
             return (self.kernel.bounds,)
 
-    if normalize is True:
-        return Normalize(ConvolutionOf(kernel))
-    else:
-        return ConvolutionOf(kernel)
+        @property
+        def minmax(self):
+            return self.kernel.minmax
+
+    return ConvolutionOf(kernel, mean=mean)